main_production_system/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
plotly>=5.0.0
streamlit>=1.0.0
python-dotenv>=0.19.0
pyyaml>=6.0
schedule>=1.1.0
requests>=2.25.0
prophet>=1.1.0
tensorflow>=2.10.0